+++ examples/Sugarscape/src/visualisation/visualisation.h
@@ -23,16 +23,16 @@ const unsigned int WINDOW_WIDTH = 512;
 const unsigned int WINDOW_HEIGHT = 512;
 
 //frustrum
-const double NEAR_CLIP = 0.1;
-const double FAR_CLIP = 500;
+const float NEAR_CLIP = 0.1f;
+const float FAR_CLIP = 500;
 
 //Circle model fidelity
 const int SPHERE_SLICES = 10;
 const int SPHERE_STACKS = 10;
-const double SPHERE_RADIUS = 1;
+const float SPHERE_RADIUS = 1;
 
 //Viewing Distance
-const double VIEW_DISTANCE = 256;
+const float VIEW_DISTANCE = 256;
 
 //light position
 GLfloat LIGHT_POSITION[] = {10.0f, 10.0f, 10.0f, 1.0f};
